+++ packages/synapse-interface/components/ui/SelectSpecificNetworkButton.tsx
@@ -28,7 +28,7 @@ export const SelectSpecificNetworkButton = ({
   const ref = useRef<any>(null)
   const chain = CHAINS_BY_ID[itemChainId]
 
-  const buttonClass = joinClassNames({
+  const buttonClasses = {
     other: 'whitespace-nowrap',
     grid: 'grid gap-0.5',
     space: 'pl-2 pr-1.5 py-2.5 w-full',
@@ -37,12 +37,12 @@ export const SelectSpecificNetworkButton = ({
     hover: getHoverStyleForButton(chain?.color),
     activeStyle:
       isActive || isSelected ? getActiveStyleForButton(chain?.color) : '',
-  })
+  }
 
   return (
     <button
       ref={ref}
-      className={buttonClass}
+      className={joinClassNames(buttonClasses)}
       onClick={onClick}
       data-test-id={`${dataId}-item`}
     >